The model is separated into a separate module.
