+++ app/build.gradle.kts
@@ -86,6 +86,9 @@ dependencies {
     //Glide
     implementation(Dependencies.GLIDE_DEP)
 
+    //Module AppCore
+    implementation(project(mapOf("path" to ":AppCore")))
+
     testImplementation("junit:junit:4.+")
     androidTestImplementation("androidx.test.ext:junit:1.1.3")
     androidTestImplementation("androidx.test.espresso:espresso-core:3.4.0")
