+++ app/src/main/java/ru/brauer/cleanarchitecture/model/datasource/RetrofitImplementation.kt
@@ -7,7 +7,7 @@ import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
-import ru.brauer.cleanarchitecture.model.data.DataModel
+import ru.brauer.appcore.model.data.DataModel
 
 class RetrofitImplementation : DataSource<List<DataModel>> {
     override fun getData(word: String): Observable<List<DataModel>> {
