+++ app/src/main/java/ru/brauer/cleanarchitecture/model/repository/RepositoryImplementation.kt
@@ -1,7 +1,7 @@
 package ru.brauer.cleanarchitecture.model.repository
 
 import io.reactivex.rxjava3.core.Observable
-import ru.brauer.cleanarchitecture.model.data.DataModel
+import ru.brauer.appcore.model.data.DataModel
 import ru.brauer.cleanarchitecture.model.datasource.DataSource
 
 class RepositoryImplementation(private val dataSource: DataSource<List<DataModel>>) :
