+++ app/src/main/java/ru/brauer/cleanarchitecture/model/repository/RepositoryLocalImplementation.kt
@@ -1,7 +1,7 @@
 package ru.brauer.cleanarchitecture.model.repository
 
 import io.reactivex.rxjava3.core.Observable
-import ru.brauer.cleanarchitecture.model.data.DataModel
+import ru.brauer.appcore.model.data.DataModel
 import ru.brauer.cleanarchitecture.model.datasource.DataSourceLocal
 
 class RepositoryLocalImplementation(private val dataSource: DataSourceLocal) : RepositoryLocal {
