+++ app/src/main/java/ru/brauer/cleanarchitecture/view/main/MainInteractor.kt
@@ -3,7 +3,7 @@ package ru.brauer.cleanarchitecture.view.main
 import io.reactivex.rxjava3.core.Observable
 import ru.brauer.cleanarchitecture.interactor.Interactor
 import ru.brauer.cleanarchitecture.model.data.AppState
-import ru.brauer.cleanarchitecture.model.data.DataModel
+import ru.brauer.appcore.model.data.DataModel
 import ru.brauer.cleanarchitecture.model.repository.Repository
 import ru.brauer.cleanarchitecture.model.repository.RepositoryLocal
