+++ settings.gradle
@@ -8,3 +8,4 @@ dependencyResolutionManagement {
 }
 rootProject.name = "Clean architecture"
 include ':app'
+include ':AppCore'
